+++ src/components/Modules/PostCard/index.ts
@@ -1,0 +1,3 @@
+import PostCard from './PostCard'
+
+export default PostCard
