+++ src/components/Modules/PostCard/styles.css.d.ts
@@ -1,0 +1,8 @@
+export const PostCard: string
+export const header: string
+export const body: string
+export const img: string
+export const date: string
+export const title: string
+export const tagList: string
+export const tag: string
